Add jsonl intermediate caching to signals. Introduce a central spot for this cache abstraction. (#858)
## High-level abstractions on `DatasetDuckDB`:
### `_compute_disk_cached`
Computes a signal, or map function, over a selected input. Every output
is saved to disk via jsonl, and can be sharded amongst multiple
processes.

### `_reshard_cache`
Re-shards the jsonl files into a parquet file with duckdb.

These two abstractions:
- Allow restarting based on the cache-key of the output path by
anti-joining to the jsonl files. We assume here that shards are
consistent (e.g. the select sorts by ROWID)
- Compute progress via the UI.
- Allow explicitly overwriting (will throw if it exists)
- Embedding computation does *not* get cached via jsonl. It get cached
as it previously did. This means if you run an embedding again, there
are no restarts. We can fix this later.

These two are now used for Dataset.map() and Dataset.compute_signal().

Other changes:
- Instead of ranges, shard indices and counts are sent to workers of map
- I removed the column selection for dataset.map(), it always selects
everything. This is kind of confusing, and and over-generalization.

Side-effects:
- There's a funky side-effect where signals now can emit float32, but on
the other side of select_rows it will be float64 because of going
through jsonl. This can be solved by explicitly passing a schema to the
COPY {json} TO PARQUET but is non-trivial because of nested structs.


As a side-effect, signals also can now have optional schemas. I didn't
do this yet, but I'll do it in a follow up to keep this smaller.

Demo: https://huggingface.co/spaces/lilacai/nikhil_staging


https://github.com/lilacai/lilac/assets/1100749/00e43943-8dbe-4d5f-ad2b-1e8c7150ecbd

+++ lilac/data/dataset_export_test.py
@@ -131,8 +131,8 @@ def test_export_to_pandas(make_test_data: TestDataMaker) -> None:
   df = dataset.to_pandas([ROWID, 'text', 'text.test_signal.flen'])
   expected_df = pd.DataFrame(
     [
-      {ROWID: '1', 'text': 'hello', 'text.test_signal.flen': np.float32(5.0)},
-      {ROWID: '2', 'text': 'everybody', 'text.test_signal.flen': np.float32(9.0)},
+      {ROWID: '1', 'text': 'hello', 'text.test_signal.flen': np.float64(5.0)},
+      {ROWID: '2', 'text': 'everybody', 'text.test_signal.flen': np.float64(9.0)},
     ]
   )
   pd.testing.assert_frame_equal(df, expected_df)
